Renamed JS file for consistency. Global css hover effect edts.
New version of portfolio-slider-js renamed to projects-page-js, since it contains general JS for the projects page and that is more consistent with the file naming conventions for other pages.

Updated link-hover-ix-02 to invert images, so all <img> children will be inverted on hover. Needed for logo hover animation change and easier than styling per element.